+++ __tests__/fixtures/rules/jsx-no-duplicate-props.tsx
@@ -1,2 +1,0 @@
-// @expect-rule: smartm2m-react/no-duplicate-jsx-props
-export const DuplicateProps = () => <button disabled={false} disabled />;